requirements.txt
langchain-groq
mysql-connector-python
pandas
gradio
python-dotenv
psycopg2
SQLAlchemy